# packages/api-client/src/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type Body_login_access_token_api_v1_auth_login_post = {
    grant_type?: string | null;
    username: string;
    password: string;
    scope?: string;
    client_id?: string | null;
    client_secret?: string | null;
};

export type ClientCreate = {
    email: string;
    full_name: string;
    password?: string | null;
};

export type ExerciseSetCreate = {
    exercise_id: number;
    reps: number;
    weight_kg: number;
    order: number;
    is_completed?: boolean;
};

export type HTTPValidationError = {
    detail?: Array<ValidationError>;
};

export type ProfileOut = {
    full_name?: string | null;
    avatar_url?: string | null;
};

export type Tenant = {
    name: string;
    slug: string;
    is_active?: boolean;
    id?: number | null;
    created_at?: string;
    updated_at?: string;
};

export type Token = {
    access_token: string;
    refresh_token: string;
    token_type: string;
};

export type UserOut = {
    email: string;
    is_active?: boolean;
    is_superuser?: boolean;
    id: number;
    profile?: ProfileOut | null;
};

export type ValidationError = {
    loc: Array<(string | number)>;
    msg: string;
    type: string;
    input?: unknown;
    ctx?: {
        [key: string]: unknown;
    };
};

export type WorkoutCreate = {
    name: string;
    intensity: string;
    volume: string;
    duration_minutes: number;
    client_id: number;
    exercises?: Array<ExerciseSetCreate>;
};

export type WorkoutOut = {
    name: string;
    intensity: string;
    volume: string;
    duration_minutes: number;
    client_id: number;
    id: number;
    completed: boolean;
    created_at: string;
};

export type ReadUserMeApiV1AuthMeGetResponse = UserOut;

export type ReadUserMeApiV1AuthMeGetError = unknown;

export type LoginAccessTokenApiV1AuthLoginPostData = unknown;

export type LoginAccessTokenApiV1AuthLoginPostResponse = Token;

export type LoginAccessTokenApiV1AuthLoginPostError = unknown;

export type ReadTenantsApiV1TenantsGetResponse = Array<Tenant>;

export type ReadTenantsApiV1TenantsGetError = unknown;

export type CreateClientApiV1ClientsPostData = {
    body: ClientCreate;
    query?: {
        tenant_id?: number | null;
    };
};

export type CreateClientApiV1ClientsPostResponse = UserOut;

export type CreateClientApiV1ClientsPostError = unknown;

export type ListClientsApiV1ClientsGetData = {
    query?: {
        tenant_id?: number | null;
    };
};

export type ListClientsApiV1ClientsGetResponse = Array<UserOut>;

export type ListClientsApiV1ClientsGetError = unknown;

export type CreateWorkoutApiV1WorkoutsPostData = {
    body: WorkoutCreate;
    query?: {
        tenant_id?: number | null;
    };
};

export type CreateWorkoutApiV1WorkoutsPostResponse = WorkoutOut;

export type CreateWorkoutApiV1WorkoutsPostError = unknown;

export type ListWorkoutsApiV1WorkoutsGetData = {
    query?: {
        tenant_id?: number | null;
    };
};

export type ListWorkoutsApiV1WorkoutsGetResponse = Array<WorkoutOut>;

export type ListWorkoutsApiV1WorkoutsGetError = unknown;

export type ListMetricsApiV1MetricsGetResponse = unknown;

export type ListMetricsApiV1MetricsGetError = unknown;

export type GetNutritionApiV1NutritionGetResponse = unknown;

export type GetNutritionApiV1NutritionGetError = unknown;

export type GenerateWorkoutApiV1AiGenerateWorkoutPostResponse = unknown;

export type GenerateWorkoutApiV1AiGenerateWorkoutPostError = unknown;

export type AnalyzeMetricsApiV1AiAnalyzeMetricsPostResponse = unknown;

export type AnalyzeMetricsApiV1AiAnalyzeMetricsPostError = unknown;

export type ChatbotApiV1AiChatPostResponse = unknown;

export type ChatbotApiV1AiChatPostError = unknown;

export type ListIntegrationsApiV1IntegrationsGetResponse = unknown;

export type ListIntegrationsApiV1IntegrationsGetError = unknown;

export type ReadRootGetResponse = unknown;

export type ReadRootGetError = unknown;

export type $OpenApiTs = {
    '/api/v1/auth/me': {
        get: {
            res: {
                /**
                 * Successful Response
                 */
                '200': UserOut;
            };
        };
    };
    '/api/v1/auth/login': {
        post: {
            req: LoginAccessTokenApiV1AuthLoginPostData;
            res: {
                /**
                 * Successful Response
                 */
                '200': Token;
                /**
                 * Validation Error
                 */
                '422': HTTPValidationError;
            };
        };
    };
    '/api/v1/tenants/': {
        get: {
            res: {
                /**
                 * Successful Response
                 */
                '200': Array<Tenant>;
            };
        };
    };
    '/api/v1/clients/': {
        post: {
            req: CreateClientApiV1ClientsPostData;
            res: {
                /**
                 * Successful Response
                 */
                '200': UserOut;
                /**
                 * Validation Error
                 */
                '422': HTTPValidationError;
            };
        };
        get: {
            req: ListClientsApiV1ClientsGetData;
            res: {
                /**
                 * Successful Response
                 */
                '200': Array<UserOut>;
                /**
                 * Validation Error
                 */
                '422': HTTPValidationError;
            };
        };
    };
    '/api/v1/workouts/': {
        post: {
            req: CreateWorkoutApiV1WorkoutsPostData;
            res: {
                /**
                 * Successful Response
                 */
                '200': WorkoutOut;
                /**
                 * Validation Error
                 */
                '422': HTTPValidationError;
            };
        };
        get: {
            req: ListWorkoutsApiV1WorkoutsGetData;
            res: {
                /**
                 * Successful Response
                 */
                '200': Array<WorkoutOut>;
                /**
                 * Validation Error
                 */
                '422': HTTPValidationError;
            };
        };
    };
    '/api/v1/metrics/': {
        get: {
            res: {
                /**
                 * Successful Response
                 */
                '200': unknown;
            };
        };
    };
    '/api/v1/nutrition/': {
        get: {
            res: {
                /**
                 * Successful Response
                 */
                '200': unknown;
            };
        };
    };
    '/api/v1/ai/generate-workout': {
        post: {
            res: {
                /**
                 * Successful Response
                 */
                '200': unknown;
            };
        };
    };
    '/api/v1/ai/analyze-metrics': {
        post: {
            res: {
                /**
                 * Successful Response
                 */
                '200': unknown;
            };
        };
    };
    '/api/v1/ai/chat': {
        post: {
            res: {
                /**
                 * Successful Response
                 */
                '200': unknown;
            };
        };
    };
    '/api/v1/integrations/': {
        get: {
            res: {
                /**
                 * Successful Response
                 */
                '200': unknown;
            };
        };
    };
    '/': {
        get: {
            res: {
                /**
                 * Successful Response
                 */
                '200': unknown;
            };
        };
    };
};